move 2.35 fix to api requests
